backend/productcomparator/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
gunicorn==21.2.0
requests==2.31.0
firebase-admin==6.6.0
python-dotenv==1.0.1
pydantic==2.0.0
baml-py==0.83.0